% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key.R
\name{get_key}
\alias{get_key}
\title{Get Key}
\usage{
get_key(url, username, password)
}
\arguments{
\item{url}{URL of the BD API gateway}

\item{username}{user name for BrownDog}

\item{password}{password for BrownDog}
}
\value{
BD API key
}
\description{
Get a key from the BD API gateway to access BD services
}
\examples{
\dontrun{
get_key("https://bd-api-dev.ncsa.illinois.edu", "your email", "password")
}
}
