% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_formats.R
\name{get_output_formats}
\alias{get_output_formats}
\title{Get input format.}
\usage{
get_output_formats(url, inputformat, token)
}
\arguments{
\item{url}{The URL to the Brown Dog server to use.}

\item{inputformat}{The format of the input file.}

\item{token}{Brown Dog access token}
}
\value{
: A string array of reachable output format extensions.
}
\description{
Check Brown Dog Service for available output formats for the given input format.
}
\examples{
\dontrun{
key <- get_key("https://bd-api-dev.ncsa.illinois.edu", "your email", "password")
token <- get_token("https://bd-api-dev.ncsa.illinois.edu", key)
get_output_formats("https://bd-api-dev.ncsa.illinois.edu", "csv", 
token)
}
}
