% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get Token}
\usage{
get_token(url, key)
}
\arguments{
\item{url}{URL of the BD API gateway}

\item{key}{permanet key for BD API}
}
\value{
BD API Token
}
\description{
Get a Token from the BD API gateway to access BD services
}
\examples{
\dontrun{
key <- get_key("https://bd-api-dev.ncsa.illinois.edu", "your email", "password")
get_token("https://bd-api-dev.ncsa.illinois.edu", key)
}
}
